requirements.txt
aiogram
requests
pillow
pytesseract
torch
numpy
pydub
transformers